Abstract read buffer creation
