+++ src/Storages/ObjectStorage/DataLakes/Iceberg/IcebergMetadata.cpp
@@ -41,6 +41,8 @@
 #include <Storages/ObjectStorage/DataLakes/Iceberg/PositionDeleteTransform.h>
 #include <Storages/ObjectStorage/DataLakes/Iceberg/Constant.h>
 
+#include <Storages/ObjectStorage/Utils.h>
+
 #include <Common/logger_useful.h>
 #include <Common/ProfileEvents.h>
 #include <Common/SharedLockGuard.h>
@@ -549,7 +551,7 @@ ManifestFileCacheKeys IcebergMetadata::getManifestList(ContextPtr local_context,
         if (manifest_cache)
             read_settings.enable_filesystem_cache = false;
 
-        auto manifest_list_buf = StorageObjectStorageSource::createReadBuffer(object_info, object_storage, local_context, log, read_settings);
+        auto manifest_list_buf = createReadBuffer(object_info, object_storage, local_context, log, read_settings);
         AvroForIcebergDeserializer manifest_list_deserializer(std::move(manifest_list_buf), filename, getFormatSettings(local_context));
 
         ManifestFileCacheKeys manifest_file_cache_keys;
@@ -691,7 +693,7 @@ ManifestFilePtr IcebergMetadata::getManifestFile(
         if (manifest_cache)
             read_settings.enable_filesystem_cache = false;
 
-        auto buffer = StorageObjectStorageSource::createReadBuffer(manifest_object_info, object_storage, local_context, log, read_settings);
+        auto buffer = createReadBuffer(manifest_object_info, object_storage, local_context, log, read_settings);
         AvroForIcebergDeserializer manifest_file_deserializer(std::move(buffer), filename, getFormatSettings(local_context));
 
         return std::make_shared<ManifestFileContent>(
