+++ src/Storages/ObjectStorage/StorageObjectStorageSource.h
@@ -67,12 +67,6 @@ public:
         const ObjectInfo & object_info,
         bool include_connection_info = true);
 
-    static std::unique_ptr<ReadBufferFromFileBase> createReadBuffer(
-        ObjectInfo & object_info,
-        const ObjectStoragePtr & object_storage,
-        const ContextPtr & context_,
-        const LoggerPtr & log,
-        const std::optional<ReadSettings> & read_settings = std::nullopt);
 
 protected:
     const String name;
